Refactor entities, include mapper from Entity to Response.

+++ src/main/java/com/gbm/challenge/gbmchallenge/model/request/CreateAccountDto.java
@@ -3,9 +3,11 @@ package com.gbm.challenge.gbmchallenge.model.request;
 import lombok.Getter;
 import lombok.NonNull;
 
+import java.math.BigDecimal;
+
 @Getter
 public class CreateAccountDto {
 
-    @NonNull private Long cash;
+    @NonNull private BigDecimal cash;
 
 }
